postgres_to_es/requirements.txt
backoff==1.10.0
psycopg2-binary==2.8.6
pydantic==1.8.2
python-dotenv==0.17.1
requests==2.25.1
